refactor: extract update actions from App
